+++ docs/changes.rst
@@ -6,6 +6,10 @@ Version 0.4.0
 
 To be released.
 
+- Now master key can be found without token through
+  :http:get:`/masterkey/` API.
+- Deprecated :http:get:`/tokens/(token_id:token_id)/masterkey/`.
+  Use :http:get:`/masterkey/` instead.
 - Geofront becomes to require Paramiko 2.0.0 or higher.
 - Geofront becomes to require Werkzeug 0.11 or higher.
 - Geofront becomes to require Flask 0.10.1 or higher.